requirements.txt
annotated-doc==0.0.4
anyio==4.13.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
cssselect==1.4.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.29.0
Flask==3.1.3
flask-cors==6.0.2
fsspec==2026.4.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.13.0
idna==3.13
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.3
lxml==6.1.0
lxml_html_clean==0.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
newspaper3k==0.2.8
nltk==3.9.4
numpy==2.4.4
packaging==26.2
pillow==12.2.0
Pygments==2.20.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-file==3.0.1
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
skorch==1.3.1
soupsieve==2.8.3
sympy==1.14.0
tabulate==0.10.0
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.1
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.7.0
typer==0.25.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
